Use temporary mesh to accumulate gparts to hashmap
When assigning gparts from a cell to the hashmap we allocate
a small patch of mesh covering the cell, accumulate the gparts
to this mesh, then copy all the values from the mesh to the
hashmap. This should reduce the number of hashmap lookups needed
and makes it easier to parallelize over cells.