Avoid rounding error issue in MPI mesh gravity
In mpi_mesh_fetch_potential() we have to fetch the potential mesh
cells that will later be needed to compute accelerations. It seems
that rounding error can cause us to fail to import cells which we
need if a particle is right on the cell boundary. Added a buffer
of 1% of the mesh cell size to prevent crashes.

Previously this problem was masked by the extra boundary added to
allow for particles moving since the mesh was created.